requirements.txt
beautifulsoup4==4.13.4
cassandra-driver==3.29.3
neo4j==5.28.1
pymongo==4.12.0
python-dotenv==1.1.0
requests==2.32.3
